rename three sixty to panorama
