+++ vegbilder-client/src/components/ImageView/ImageViewer/PanoramaImage/PanoramaImage.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactPannellum from "react-pannellum";
 import Theme from "theme/Theme";
 import testImg from './test-img.png';
-import './style.css';
+import './panellumStyle.css';
 
 
 const ThreeSixtyImage = () => {
@@ -18,13 +18,6 @@ const ThreeSixtyImage = () => {
     showFullscreenCtrl: false,
     uiText: uiText,
   };
-
-  // const authorBox = document.getElementsByClassName('pnlm-panorama-info').item(0);
-
-  // if (authorBox){
-  //   authorBox.style['display'] = 'none';
-  // }
-   
 
     return (
       <>
