+++ vegbilder-client/src/components/ImageView/ImageViewer/PanoramaImage/panellumStyle.css
